safari-600.7-branch/LayoutTests/editing/selection/DOMSelection-DocumentType-expected.txt

Test to check if setBaseAndExtent guard node with null owner document (Bug 31680)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.anchorNode is null
PASS sel.containsNode(docType) is false
PASS successfullyParsed is true

TEST COMPLETE
